Remove movist-v2-default-input.conf from Copy Bundle Resources
This commit will remove the file movist-v2-default-input.conf from the
Copy Bundle Resources build phase.

The default key bindings configuration files are stored in a config
subdirectory of Resources. These files are copied in the Copy Configs
build phase. The movist-v2-default-input.conf file was included in both
build phases resulting in two copies of the file being in the
application bundle. Only the file in the Resources/config directory is
used by IINA. This change removes the duplicate in the Resources
directory.