Fix OSC next/prev media buttons skip files, #5611 (#5763)
This commit will:
- Rename the updateArrowButtonImage method in MainWindowController to be
  updateArrowButtons
- Change updateArrowButtons to configure the button type to be
  momentaryPushIn or multiLevelAccelerator based on what the OSC buttons
  are set to do
- Change the leftButtonAction and rightButtonAction methods to handle
  momentaryPushIn buttons

With these changes the OSC left and right arrow buttons only change the
position in the playlist by one file. This also corrects the behavior
of the buttons when set to seek to only seek once.

This also corrects the behavior where a pressure sensitive device was
providing tactile feedback when the buttons were configured to change
the position in the playlist, or seek forward. As the buttons only take
into account the pressure being applied when the buttons are set to
control playback speed it was inappropriate to be providing pressure
feedback when the buttons were not controlling the speed.